Use Grunt instead of make
Better support for developing on Windows, and leverages Grunt's large ecosystem
of plugins.
