+++ package.json
@@ -2,12 +2,8 @@
   "name": "metalsmith.io",
   "private": true,
   "dependencies": {
-    "async": "^0.2.10",
     "component": "^0.19.9",
     "component-builder": "~0.11.2",
-    "consolidate": "~0.10.0",
-    "express": "~3.4.8",
-    "extend": "^1.2.1",
     "metalsmith": "^1.7.0",
     "metalsmith-in-place": "^1.3.1",
     "metalsmith-layouts": "^1.4.1",
@@ -15,6 +11,15 @@
     "metalsmith-metadata": "0.0.1",
     "mkdirp": "^0.3.5",
     "myth": "^0.1.0",
-    "swig": "^1.3.2"
+    "swig": "^1.3.2",
+    "grunt": "^1.0.1",
+    "grunt-contrib-clean": "^1.0.0",
+    "grunt-contrib-connect": "^1.0.2",
+    "grunt-shell": "^1.3.1"
+  },
+  "scripts": {
+    "build": "grunt build",
+    "prepublish": "npm run build",
+    "start": "grunt connect"
   }
 }
